finally -> ensure
Since `report` (was `catch`) is terminating, `finally` is named such that it implies you should add this handler last, but you can’t. Also since we can no longer use `catch`, `finally` is legacy and not related to the pattern anymore, so instead we will use `ensure`.

I decided to leave AnyPromise with `catch` and `finally`. For one, such change is anti-consumer, and for two, the AnyPromise `catch` behaves differently (since it is also `recover`) and they can be written more like `@catch` and `@finally`.

Really, `ensure` is a better name for promises, often you don’t want that handler to be the final handler.

I played with `undo` but it had too many other connotations even though it looked good in code samples.